This allows for a
{{{
<include file="fileinsamedirasmaster.conf"/>
}}}

tp be placed anywhere in the master config of either noitd or stratcond.
The include is processed by parsing the specified file and putting all
children of the root element of the included document as replacements for
the children of the {{{<include>}}} node itself.

Special care is taken to make it look like we did not do this when we go to
write the config anywhere.

refs #254