Stop composite graphs throwing errors on nil value
Currently if you are creating a composite graph and the numeric value is
nil, you will get an error and the composite won't be drawn. This can
happen when you're graphing a range that doesn't have data yet (e.g.
looking at a 6 month view with only 4 months of data), or it can happen
if you've not collected a value (e.g. in our case collectd has sent a
  nil value or crashed and not reported a value for a timeframe).

This change will check to see if you are graphing numeric data, if it
finds a nil value in the results array it will replace it with a zero.
This allows the composite to still be drawn despite the missing data.