Handle nil values gracefully in composites
Currently the composite graph functions will throw an error if you're
trying to create a composite graph and there are nil values in the
numeric data source.

This patch aims to fix that by returning the nil values as a null, then
treating them as 0.0 when it comes to calculating the graph.