Allow checks to note that they are passive.
If a check doesn't "do" anything to collect metrics and instead
passively allows metrics to arrive, then when you clone the check
to a transient copy, the metrics will always be blank.  By
allowing modules to note that the check acts this way, we can
be smarter elsewhere.  We should be able to use this to allow
transient streaming of things like statsd, collectd and traps.