Make failed PCREs behave consistently.
If a pcre failed to compile, it would be as if it wasn't specified.
This causes it to "always match" which is quite unintuitive.  Now,
if the pcre isn't specified, that behavior remains, but if a PCRE
fails to compile, we establish a fallback pcre that never matches
anything: ^(?=a)b

This has a behavior chaning effect for those relying on broken PCREs
in filtersets.