Fix NRPE message and state metrics
State can have spaces in it like DISK OK, so allow them and then
match an OK in it to have a good status

Message should pull everything up to a | i.e. until it encounters
perf data

Also changed the regex to only pull the first line of data for now
as a stop gap until something better is produced.  At the moment
multi line nrpe breaks it horribly anyway so this is no less
incorrect.