Change Epoll Error To Debug Level Message
When we reach the condition where epoll tries to modify a descheduled
fd, nothing actually breaks, but we can possibly fill the log with
error messages. Since this isn't crucial, but we may still want to be
aware of it for debugging purposes, changed the level of the message
from noit_error to noit_debug.