Remove Unneeded Debug Message, More Precise rlimit_nofile Setting
The noit watchdog was timing out the rlimit setting code when being
run with debugging on, which was spamming a ton of debug messages
to the log. Remove this message and go back to going back by one,
trying to set every rlimit value until the correct one is found to
increase precision of setting the maximum possible open file
descriptors.