Add Ability To Spawn Multiple DNS Contexts In DNS Check
There was an issue where if a DNS check was being run with a very
high frequency and the DNS server was not responding, we could end
up with long backups in the UDNS, causing watchdog timeouts. Added
the ability to split into multiple contexts per namespace, reducing
the load of DNS queries that any one will have to deal with.