Pull connection creation into DB specific modules, support Oracle RAC
In order to support Oracle RAC and their fast connection failover, we
need to use a different class to get the DB connection rather than
DriverManager.

To accomplish this, pulled the connection creation via getConnection
into the specific classes, and modified Oracle check to use reflection
to pull in the OracleDataSource class if we have it, otherwise use
DriverManager.

RAC is supported by using the connect string as your database param,
we build the connect url by then using just that string.