Don't Cache Cloudwatch Data For Longer Than A Minute
We were allowing Cloudwatch caches to be up to five minutes long.
This could have a negative impact if users are configured to pull
five-minute granularity data, but are receiving one-minute updates.
Changed the cache to always hold data for about a minute (55
seconds to allow a little wiggle room).