Add Hack To Verify RabbitMQ End Byte
Added a hacky check to verify that the end byte for RabbitMQ body
messages is indeed 206. We're seeing issues where this is coming
across incorrectly and we need to verify if somehow, things are
getting set incorrectly on the Reconnoiter end. Once figured out,
this should probably be taken out.