Add "Round Robin" Option To FQ Message Publishing
Added an option to go "round robin" when publishing FQ messages;
this will publish the data in a circle around the available FQs
rather than sending all messages to all available FQs.

If a node is determined to be down (by getting a "socket: Connection
refused" message from FQ), we will try to send to other good nodes
instead. We will publish to all nodes that are down to minimize
possible data loss.

This is experimental and will not guarantee that all messages will
be received, so use at your own risk.