Add API Call To Add A Jlog Subscribed
Added an API call to allow creating a new Jlog subscriber. This is a PUT
call that takes an XML object that contains the name of the new
subscriber and the existing subscriber to copy the checkpoint from. If
this is not provided, it sets the checkpoint to JLOG_END.