Add Filtered Exchange To Stratcon
Added the ability to have a "filtered" exchange in Stratcon. This will
only pass certain checks and metrics through. The checks and/or
metrics required can be specified in XML via a PUT command, which will
then pass the information on to the RabbitMQ and FQ drivers.

Right now, the RabbitMQ driver is not implemented; only the FQ driver
is.