Add "minutes_back" Parameter To Cloudwatch Check
Added a new parameter to the Cloudwatch check, "minutes back". This
tells the check the number of minutes back from the present time to scan
when looking for metrics. The check will always take the most recent
metric it finds in that range as the "official" value.

Change the default number of minutes to read back to 10.