Fix last_fire_time on httptrap checks.
We have now, just use that.  This was causing a double up
of stats sets, which cause the check to skip firing every
other time it was run.  With asynch_metrics on, this didn't
hurt anything, but the histogram module can inject metrics
and those were not getting logged every other time.