Use mtevFatal Instead Of assert/abort
Replaces tons of assert/abort calls with calls to mtevFatal, which logs
synchronously before exiting. This is better for debugging - it makes
sure that we get errors in logs before we abort/assert out.

I didn't replace every single instance, since there are tons of them,
but I replaced a bunch of the more important ones.