memchr instead of memmem.
I'll take a timeout and drop a reminder that Linux is one raging tire fire
after another.  An entire generation of software engineer will or already
have the equivalent of coal miner's blacklung from this tragic and
asphyxiating occupational hazard.  The viral nature of bad technical
practices that have metastasized into a cancer is poised to communicate
to the next generation of systems programmers.  I see no cure.  One
day soon programmers will be coughing that sickening, saddening blacklung
cough and finally die.  Hopeless and alone with a legacy of malfunctioning
software.

equivalent