Set "null" cloudwatch values to 0
Cloudwatch is super annoying in that if they don't have a value for
something, they simply don't return it.

I haven't looked at every single metric, but I think in most every
case this would mean the value is 0 (like requests, etc) because if
there was a value, it would be returned.

So, when we get back a datapoints section that is empty, loop over
the stats we requested, and return a 0 value instead of nothing for
that metric