Ensure timeouts in DNS checks are thread-local to DNS fds.
We were casualy about creating timeouts for DNS checks.  The check
would fire on some thread and create a timeout on that thread which
could be different from the thread managing the DNS server connection
socket.  This could cause races where DNS completect while the timeout
was firing and vice versa.  By making the timeouts fire on the same
thread that owns the socket, we avoid this race condition.