Add newer DH groups from RFC 3526
Rename the 2048-bit group as group14, which it is, and to be consistent
with the naming of the larger groups.

Add a TODO comment about supporting RFC 4419 group exchange, where
there is no fixed group and the server picks a group that closely
matches client preference. We will need to implement the logic for
this.