Modernize lua ssh2 module (#405)
* Add SHA256 fingerprint support

Also add base64-encoded versions of the sha1 and sha256 fingerprint
metrics. This matches recent OpenSSH (>= 6.8) behavior, which is to
show base64-encoded strings rather than hex-encoded.

* Add host key types ecdsa-sha2-nistp256 and ssh-ed25519

* Add newer DH groups from RFC 3526

Rename the 2048-bit group as group14, which it is, and to be consistent
with the naming of the larger groups.

Add a TODO comment about supporting RFC 4419 group exchange, where
there is no fixed group and the server picks a group that closely
matches client preference. We will need to implement the logic for
this.

* Update cipher list

Add AES-GCM and ChaCha20.

* Update docs for cipher changes

Add to the XML description any options that weren't there originally,
such that we maintain agreement between code and documentation.

* Add newer MAC algorithms