Add Ability To Test For And Unlink Trash Files In Stratcon Sweep
When sweeping through stratcon files, there are occasionally trash files
in the directory that we ignore, but we really need to clean up.

Added a new paramter to the sweep function, "unlink_test", which is a
function that determines if we should unlink an invalid file. If this
returns true, we will attempt to unlink the file.

This is currently hooked up to unlink a valid raw_ingestion file with a
".tmp" extension at the end.