Fix missing namespace detection.
It was broken, this is wasterful, now it shoudl work well enough.
Namespaces in XML docs are interned, so pointer comparison can
be used (per libxml2 docs). We also expect a small set of name
spaces for checks XML.  So, we keep an array of namespaces we've
seen, brute force searching and appending when we see a new one.
We recursively search the src document for namespaces to add.
Then we search for each one once in the tgt document.