Implement //checks/@perpetual_metrics (#479)
Change the default behavior of check stats to not be perpetually
cummulative (and thus potentially memory exhausting).  The old
behavior can be renabled by setting the option to true; it defaults
false.

The old behavior was that current stats were replaced into previous
stats by name (so old metrics stuck around perpetually).  The new
default mirrors the original behavior in that the current stats
simply become the previous stats.