XML backed *_auto_add rules are untenable. (#501)
* XML backed *_auto_add rules are untenable.

The XML config change rate is too high and cannot be maintained
reasonable using libxml2.  Since we're trying to manage concurrency
more than total historic unique metrics, it makes sense to only
manage these hashes transiently.  Additionally, as metrics come and
go, it makes sense to refresh the automatic set periodically. We
default this refresh to 300000ms (5 minutes) and allow it to be
set via filter_flush_period (0 to disable).