mkversion.sh should detect detached HEAD state
The value returned by `git name-rev` when in detached-HEAD is
unpredictable and of low value, since it will pick the nearest symbolic
name in lexicographic order, which really does nothing for
communicating "what branch we're on".

Instead, use `git symbolic-ref` to resolve HEAD, and if it's empty,
we must be detached and should just say so.

This solution also works around the lack of a `--short` option to
symbolic-ref in older git versions, so we will still get "master"
instead of "refs/heads/master", for example.