Bundled immediate trap (#528)
* Make httptrap immediate mode more efficient.

Immediate mode would cause individual log lines to be emitted,
but as B1 records which has overhead that compression doesn't
work well on with only a few records.  The trap now accrues
metrics that should be logged immediately and batch/bundles them
in larger B1 payloads.  This changes from a "negative" compression
rate on single recoreds to 60-80% compression rate on batches of
100 or more.

* Remove error log to debug

* disable metric director until it is requested

* address error checking