Lock around hash manipulation.
These are unfortunately heavyweight locks but the mtev_hash here
is passed into the logging system and isn't aware of memory safe
hashes where we could spmc iterate.

Adding a new API to the logging system that uses ck_hs directly
might be a better approach, but this should add much needed safety
in the interim.