CIRC-4638 - Add LMDB Backing Option For Checks (#597)
Added a flag to the "checks" configuration - "use_lmdb" - that, if set to true,
will enable an LMDB db for storing checks instead of XML. This will require setting
an "lmdb_path" field as well to tell it where the LMDB database will be stored. In
addition, there is a true/false flag, "convert_xml_to_lmdb", that will convert the existing
XML backingstore to use LMDB.

LMDB reads/writes are considerably faster than XML configuration - this path will enable
a significant speed-up when reading/writing checks - particularly noticeable if there are
a considerable number of checks.