CIRC-5074 - Use Conf Sync Flag In Cluster Busted Tests
There's a pending change in libmtev that uses a header to force cluster
puts to write out the configuration file immediately. Add a header to
the tests that use this so that when that change goes live, we'll use
it. It will have no effect until the libmtev change lands.