noit_metric_director: Separate counts for threshold/backlog
With this change mtev stats will contain the following entries:

```
$ curl -s 'localhost:8081/mtev/stats.json?format=simple' | jq '.noit'
{
  "metric_director": {
    "distributed": 3578,
    "delivered": 32009,
    "queued": 32181,
    "seen": 3845,
    "dropped": {
      "too_full": 0,
      "too_late": 0
    }
  }
}
```

```
$ curl -s 'localhost:8081/mtev/stats.json?format=tagged' | jq '.' | grep dropped
  "dropped|ST[app:noit,reason:too_full,subsystem:metric_director,units:messages]": {
  "dropped|ST[app:noit,reason:too_late,subsystem:metric_director,units:messages]": {
```

The tag reason:too_full/too_late was used, so that the meaning is more clear to the operator.
Dropped reason:threshold/backlogged would be more consistent with the variable names,
but confusing to people not familiar with the code.