CIRC-5150 - Fix Race In Check Recycle Bin
The check recycle bin - where we clean up checks that have already been
deleted, but not removed from the disk - was encountering a race where
we could potentially accidentally delete a check we're attempting to
restore.

When cleaning the recycle bin, check the config to make sure we're still
set for deletion. If we're not - don't delete it.