Use New Field For Expanded Metric Names (#780)
Rather than attempting to rewrite the metric names in place if we apply a measurement tag via a filterset, which is extremely messy, just store the expanded name in a new field in the metric_t. Then, when we try to log a metric, pull the expanded name before we do so.