compile and study search AST on-demand.
Add parse_lazy and clone_lazy functions to avoid PCRE compilation during
parsing.  Do not study PCREs by default, track executions and once we
have 50 calls to pcre_exec then study the PCRE.  Rework the lfu such
that the token is tracked with the returned object and is released once
the node is freed.

CIRC-8958