CIRC-9495 - fix handling of tags that are too long
`noit_metric_tagset_builder_add_many` needs access to `tag.total_size` to correctly iterate through `tagstr`.
this fixes a bug where, when a `tagnm` was `toolong` the `total_size` was not being set and would result in an overflow when trying to subtract an unknown value from the `size_t tagstr_len` which is unsigned.
add test for tag pairs that are too long.