avoid needing to define an additional value for `noit_metric_tagset_class_t`.
add implicit tags to existing tagsets, but increase the max size to allow for them.

we can increase the max size because these tagsets are copied locally and the extended size does not conflict with the functions performed on the tagset in this scope.

since it is a copy, the implicit tags are not being added to the id's actual check and stream tags.