Always Canonicalize Prometheus Metric Names
There was a comment that there was no need to canonicalize metric names
here because the system would handle it automatically; however, this is
a library function, which means we do not know how the caller will use
the data. We need to canonicalize the metric name here so we know that
the caller gets the correct name.